Movement Fixes
Fixed two issues I found with the double jump.  The character was not gaining height when double jumping with negative speed, and the character was not able to double jump when running off a platform without using the grounded jump.  I also prevented the character from flipping the sprite mid-air.