Halfway done with Camera(stuck)
I got up to Cinemachine Confiner Extension, but then when I try to create a bounding shape, the player starts to fall through the map if I create another object called CameraBackground and do the steps there. It isn't working when I try to do the steps in the Background tilemap.