Added Air Dash
Added code for a short air dash.  The idea is for it to extend the area your aerial hitboxes cover when you move, although I am not sure what that is going to look like when we add that.  You will not be invincible when air dashing, but you will be able to dodge attacks with the readjustment.  I want to go back and make the movement feel less sluggish, but I do not think I am going to add any other movement options until we get the rest of the game mostly finished.  I temporarily mapped the input to 'E' btw.