Attempt to add animations over network
Put animations in a separate script titled 'NetworkAnimator'. These animation functions are called from within the player movement script. This should allow for animations over the network because the player movement script returns on update function if the current user is not the owner of the character ensuring that you can only control your character. This has the side effect that the animations are never set for any character that is not yours. If the animations are in a different script which can be ran for each character even if it is not yours it should set those animations regardless.