Fixed uiRoot bug
When going to character select screen, the OnlineOptionsSceneManager would find the UI panel, but obv there is none of the children it expected to be there since its not in online options scene, and so an error would appear when trying to access the game object. Simply changed the if statement to check first if the children could be found, and then tried to access game object.