Possible janky fix
If this works, then the issue is that currPlayer.Get.. for the online movement script grabs a script that is different than the one being updated when the player moves. (Only for the client who is not host). Now it simply just passes one of the values which is needed. Slightly confusing because the same get method works for the blocking bool value? Perhaps because this value is accessed in target RPC as opposed to a server rpc like the s_LightAttack is